fix: 更新gorm版本

+++ models/dto/pageInfo.go
@@ -1,7 +1,7 @@
 package dto
 
 type PageInfo struct {
-	Size  uint        `json:"size"`
-	Total uint        `json:"total"`
+	Size  int64       `json:"size"`
+	Total int64       `json:"total"`
 	List  interface{} `json:"list"`
 }
